fix: Elements validation 중복제거, Elements 생성자 문자열 파라미터로 받도록 수정
